Update Structure.h
credit et debit  type long double au lieu de double

+++ Structure.h
@@ -2,6 +2,6 @@
 typedef struct Operation{
     int ID;                 // Identifiant unique de l'opération
     char Libelle[100];      // Libellé descriptif (100 caractères max)
-    double Credit;          // Montant crédité
-    double Debit;           // Montant débité
+    long double Credit;     // Montant crédité
+    long double Debit;      // Montant débité
 } Operation;
